requirements.txt
discord.py==2.3.2
PyNaCl==1.5.0
faster-whisper==0.10.0
python-dotenv==1.0.0
llama-cpp-python==0.2.16
SQLAlchemy==2.0.25
asyncio==3.4.3
numpy==1.26.1
wave==0.0.2
setuptools==68.2.2
matplotlib==3.8.0 